Add ScatterSevice

+++ src/main/java/com/kakao/kakaopay/dto/RequestScatterDTO.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class RequestScatterDTO {
 	private Long price;
-	private Long dividerNumber;
+	private int dividerNumber;
 }
